Fix a bug with peer bitfields
Previously, a bitfield was considered invalid even if it had a greater
length than ours. This fails to account for the fact that when
the number of pieces is not divisible by 8, there are extra padded 0s
at the end in the peer's message. Now we check if it's greater than or
equal to, and then truncate.